pyproject.toml
[project]
name = "siraal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "dotenv>=0.9.9",
    "google-genai>=1.66.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pywin32>=311",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
]
